Fixed external messaging

+++ src/ConsoleBot/Clients/ExternalMessagingClient/ExternalMessagingConfiguration.cs
@@ -8,7 +8,7 @@ public class ExternalMessagingConfiguration
     public string TelegramApiKey { get; set; }
 
     [Required]
-    public int TelegramChatId { get; set; }
+    public long TelegramChatId { get; set; }
 
     public bool ReceiveMessages { get; set; }
 }
